+++ src/api/transaction_test.go
@@ -989,7 +989,7 @@ func TestVerifyTransaction(t *testing.T) {
 
 	for _, tc := range tt {
 		t.Run(tc.name, func(t *testing.T) {
-			endpoint := "/api/v1/transaction/verify"
+			endpoint := "/api/v2/transaction/verify"
 			gateway := NewGatewayerMock()
 			gateway.On("VerifyTxnVerbose", &tc.gatewayVerifyTxnVerboseArg).Return(tc.gatewayVerifyTxnVerboseResult.Uxouts,
 				tc.gatewayVerifyTxnVerboseResult.IsTxnConfirmed, tc.gatewayVerifyTxnVerboseResult.Err)